ServiceCatalogAppRegistry - Local CF import
